+++ src/modes/login.js
@@ -5,20 +5,20 @@ registerMode({
     ID: "login",
     Init: function(con) {
         console.log("Login: Sending init to connection #"+con)
-        Com.Connections.send("What's your name? ")
+        Com.Connections.send(con, "What's your name? ")
         Com.Connections.stateChange(con, "Name")
     },
     StateName: function(con, msg) {
         console.log("Login: Got Name from connection #"+con)
-        Com.Connections.send("Slot me some credentials chummer! ")
+        Com.Connections.send(con, "Slot me some credentials chummer! ")
         Connections.stateChange(con, "Pass")
         //If not invalid name
-        Com.Connections.changeVar('name', msg.trim())
+        Com.Connections.changeVar(con, 'name', msg.trim())
     },
     StatePass: function(con, msg) {
         console.log("Login: Got Pass from connection #"+con)
         //if valid credentials 
-        Com.Connections.clearVar('name')
-        Com.Connections.changeMode('mainmenu')
+        Com.Connections.clearVar(con, 'name')
+        Com.Connections.changeMode(con, 'mainmenu')
     }
 })